Update AppButton to better work with Windows
For windows we have to execute commands differently.
The commands need to be run like:
CMD.exe /c ...
This commit enables AppButton to do that when the OS
is Windows.
Note that right now to simply start a program you will
either have to enter a command like "start firefox"
or have to specify the full path to the exe.
This might be worth revisiting in the future to make it
more user friendly. It should at least be documented
properly.