Allow more touch devices
My sample size of touchscreen devices is still of course small and I don't know
how all of them are implemented, but my current one uses a slightly different
implementation. This is an adjustment to be able to use the current one.
It should probably still work with the old one, though I'm not 100% sure of that
since I don't have it anymore. This area will probably need adjustments in the
future to cover all vendor quirks etc.