Macroboard: Implement button dragging in edit mode
Note: This also makes significant changes to the general layout of the
Macroboard and how it is saved.
It's no longer free standing buttons and saved as a Dictionary. It is now saved
as an Array and empty positions are filled with empty buttons (nulls in
layout.json). Old layouts are deprecated. There is a automatic migration to the
new layout in place for now, though that will be deleted in the future.
This change was made because it is a lot simpler and easier to maintain. The old
system was in theory more powerful and more how I imagined it to work, but it
just got a nightmare to work with. This new approach should still give enough
power to achieve all desired layouts.