Macroboard: Add better logging when debug is enabled
This introduces a second method of spawning processes, which can better log
stdout, stderr and exit codes of the process. This method however for now only
gets enabled when in debug mode because it adds a node to the scene tree for
every process which may, when a lot of these are active, have performance
implications.