SSHClient: Introduce timeout and keepalive
Before the ssh client was very susceptible to disconnect after some time and
then hanging when trying to open a channel. To circumvent this there is now a
default timeout of 1 second when opening a channel and a session. This alone
would still make it annoying to use because you on disconnect you would still
have to press a button/issue the command twice because it would only then detect
that a session isn't valid anymore. So this also introduces a keepalive_interval
of 300 seconds.