Migrate to a separate label for Config
Previously it was pretty annoying that what you wanted to show a user as the
name of a config item was also the key it was saved as. Meaning when I wanted to
change a user facing label I also needed to do a config migration.
To fix this there is now a separate label parameter for all config items.