Config: Replace clear button with default button
Instead of having a clear button for a enum or string array editor it now has a
return to default button, which IMO makes way more sense.
This also has the side effect that enums are now non nullable (-1 value), which
again I think makes way more sense.
String arrays are still nullable, but only if default value is "".