SSH: Major rework of SSHClient
This includes:
* SSHClient is now a RefCounted, because it doesn't need to be in the scene tree
to function.
* Switched to a internal ssh client, which makes error handling easier and more
cleanly separates the godot functions.
* Expose variables to godot for better configuration from gdscript
* Generated keys are now in the openssh format. While I really liked the
previous idea of storing it in a non standard format, it didn't actually work
and getting it to work was just too much effort. Plus if I want to release it as
a library I think this is actually a lot better, because it's the standard
format for keys.
* Add more docs