Actions: Remove obs test build
It was a nice idea, however it didn't actually work with PRs, because it would
only pull from the branch specified in _service and the actual PR changes
wouldn't even be tested. The fedora builds also are flaky and in general I think
it's best to just remove this, the value was questionable.