Remove OBS mention and ci build check
Since the builds were very flaky and would fail without faults from my side, I
decided it's best to remove the mention in the README and skip the ci action
that checks for a successful build. It's still submitted to OBS but for the time
being not a recommended way to use it (if it builds at some point again).